requirements.txt
bs4
selenium
pandas
webdriver_manager
psycopg2
flask
gunicorn
talisman
tabula-py

